crud methods for Map written and tested

+++ Java/src/main/java/com/bae/persistence/domain/Pose.java
@@ -8,40 +8,45 @@ import javax.persistence.Id;
 @Entity
 public class Pose {
 	@Id
-	@GeneratedValue(strategy=GenerationType.AUTO)
-	private long poseID;
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private int poseID;
 	private String poseName;
 	private String poseDifficulty;
-	
+
 	public Pose() {
 		super();
 	}
-	
-	public Pose(long poseID, String poseName, String poseDifficulty) {
+
+	public Pose(int poseID, String poseName, String poseDifficulty) {
 		super();
 		this.poseID = poseID;
 		this.poseName = poseName;
 		this.poseDifficulty = poseDifficulty;
 	}
-	//getters and setters
-	public long getPoseID() {
+
+	// getters and setters
+	public int getPoseID() {
 		return poseID;
 	}
-	public void setPoseID(long poseID) {
+
+	public void setPoseID(int poseID) {
 		this.poseID = poseID;
 	}
+
 	public String getPoseName() {
 		return poseName;
 	}
+
 	public void setPoseName(String poseName) {
 		this.poseName = poseName;
 	}
+
 	public String getPoseDifficulty() {
 		return poseDifficulty;
 	}
+
 	public void setPoseDifficulty(String poseDifficulty) {
 		this.poseDifficulty = poseDifficulty;
 	}
-	
 
 }
